+++ src/link/link.rs
@@ -1,5 +1,5 @@
 #![crate_name = "link"]
-#![allow(unstable)]
+#![feature(collections, core, old_io, old_path, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -12,8 +12,8 @@
 
 extern crate getopts;
 
-use std::io::fs::link;
-use std::path::Path;
+use std::old_io::fs::link;
+use std::old_path::Path;
 
 #[path="../common/util.rs"]
 #[macro_use]
@@ -22,7 +22,7 @@ mod util;
 static NAME : &'static str = "link";
 static VERSION : &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let opts = [
         getopts::optflag("h", "help", "display this help and exit"),
         getopts::optflag("V", "version", "output version information and exit"),
