+++ src/printenv/printenv.rs
@@ -1,5 +1,5 @@
 #![crate_name = "printenv"]
-#![allow(unstable)]
+#![feature(collections, core, old_io, os, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -16,7 +16,7 @@ extern crate getopts;
 extern crate libc;
 
 use std::os;
-use std::io::print;
+use std::old_io::print;
 
 #[path = "../common/util.rs"]
 #[macro_use]
@@ -24,7 +24,7 @@ mod util;
 
 static NAME: &'static str = "printenv";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("0", "null", "end each output line with 0 byte rather than newline"),
