+++ src/pwd/pwd.rs
@@ -1,5 +1,5 @@
 #![crate_name = "pwd"]
-#![allow(unstable)]
+#![feature(collections, core, old_io, os, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -13,7 +13,7 @@
 extern crate getopts;
 extern crate libc;
 
-use std::io::print;
+use std::old_io::print;
 
 #[path = "../common/util.rs"]
 #[macro_use]
@@ -22,7 +22,7 @@ mod util;
 static NAME: &'static str = "pwd";
 static VERSION: &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("", "help", "display this help and exit"),
