+++ src/tsort/tsort.rs
@@ -1,5 +1,5 @@
 #![crate_name = "tsort"]
-#![allow(unstable)]
+#![feature(collections, core, old_io, old_path, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -14,7 +14,7 @@
 extern crate getopts;
 extern crate libc;
 
-use std::io;
+use std::old_io as io;
 use std::collections::{HashSet, HashMap};
 
 #[path = "../common/util.rs"]
@@ -24,7 +24,7 @@ mod util;
 static NAME: &'static str = "tsort";
 static VERSION: &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let opts = [
         getopts::optflag("h", "help", "display this help and exit"),
         getopts::optflag("V", "version", "output version information and exit"),
