+++ src/unlink/unlink.rs
@@ -1,5 +1,5 @@
 #![crate_name = "unlink"]
-#![allow(unstable)]
+#![feature(collections, core, old_io, old_path, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -15,9 +15,9 @@
 extern crate getopts;
 extern crate libc;
 
-use std::io;
-use std::io::fs::{self, PathExtensions};
-use std::io::print;
+use std::old_io as io;
+use std::old_io::fs::{self, PathExtensions};
+use std::old_io::print;
 
 #[path = "../common/util.rs"]
 #[macro_use]
@@ -25,7 +25,7 @@ mod util;
 
 static NAME: &'static str = "unlink";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("h", "help", "display this help and exit"),
